Add methods for utility and helper
